Fix issues/15 and add message slide-in effect.
issues/15 Username shadows are applied when not necessary.
Fixed by allowing the user to disable this altogether by setting the
&shadow to a falsy value.

Added a "slide-in" effect to new messages, inspired by the Stream Labs
chat utility. This can be disabled by setting &msganim to a falsy value.

Both new query-string values are documented in the README.